+++ modules/point_processing.py
@@ -70,8 +70,7 @@ def closest_point(points: array_like, target: array_like) -> Tuple[ndarray, int]
     """
     distances = norm(np.subtract(points, target), axis=1)
 
-    # Wrap in int() to satisfy mypy.
-    index_closest = int(np.argmin(distances))
+    index_closest = np.argmin(distances)
 
     point_closest = points[index_closest]
 
@@ -251,8 +250,7 @@ def position_accuracy(points: ndarray, targets: ndarray, max_dist: float = 10) -
     """
     distances = norm(points - targets, axis=1)
 
-    # Wrap in float() to satisfy mypy.
-    return float(np.mean(distances <= max_dist))
+    return np.mean(distances <= max_dist)
 
 
 @require(
